edits to Testing pipelines on example subjects
I made some changes to the text here based on my experience running these procedures from this document, and based on some conversations we had during that process. The changes include:

* Making sure users aren't creating exemplar_test directories inside of existing datalad datasets
* updating wget command to pull scripts
* mentioning existence of fmriprep script for multi-session data
* making sure users are in /fmriprep/analysis before running qsub_calls.sh
* ensuring everything is properly datalad saved before running qsub_calls
* adding a note about what to do if one has to change the qsub_calls script.

Feel free to reject if you find these are not correct or were to specific to my experience and not generalizable.